disp: msm: sde: validate sspp input formats in plane atomic check
This change adds support to validate plane formats in
atomic check phase. Also, the current check is not
sufficient to distinguish between some formats like
TP10_UBWC and P010_UBWC which can result in scenarios
where unsupported formats are also detected as
supported formats. Additional checks are introduced
to make sure that this is handled properly.

Change-Id: I41ea8dc95dba93479e24a68aab0dba0a253c3eec
Signed-off-by: Charishma Jerripothula <quic_cjerripo@quicinc.com>